% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_bosc.R, R/detrend_bosc.R
\name{aggregate_bosc}
\alias{aggregate_bosc}
\alias{detrend_bosc}
\title{aggregate_bosc}
\usage{
aggregate_bosc(bosc, types = "real", levels = "ss-ga", overwrite = FALSE)

detrend_bosc(bosc)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{types}{Which data should be aggregated? choose between "real" and "surrogate" or concatenate to "real-surrogate" to perform aggregation on both.}

\item{levels}{Which levels of aggregation need to be created? Use "ss" for single subject and "ga" for grand average. Concatenate multiple levels with "-", e.g. "ss-ga"}

\item{overwrite}{whether to overwrite existing data. defaults to F.}

\item{method}{Choose between "demean", ...}
}
\value{
A BOSC-Object

A BOSC-Object
}
\description{
#' @description \code{aggregate_bosc}
Aggregates various level of data from a BOSC-Object.

\code{detrend_bosc}
Detrends time courses in an BOSC-Object.
}
\examples{
bosc = simulate_experiment(n = 10, n_timepoints = 10, n_trials = 10)
bosc = aggregate_bosc(bosc, types = "real", levels = "ss", overwrite = TRUE)


}
