% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bosc.R
\name{bosc}
\alias{bosc}
\title{bosc}
\usage{
bosc()
}
\value{
A bosc object
}
\description{
\code{bosc}
Creates an empty bosc object for further processing with boscillator functions.
}
\examples{
bosc = bosc()

}
\author{
René Michel
}
