% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bosc_import.R
\name{bosc_import}
\alias{bosc_import}
\title{bosc_import}
\usage{
bosc_import(
  data,
  sfreq,
  n_timepoints,
  vars = c("subj", "time", "trial", "resp"),
  level = "single_trial",
  aggregate = T
)
}
\arguments{
\item{data}{A dataframe with dense sampling data}

\item{sfreq}{in Hz}

\item{n_timepoints}{Number of unique timepoints in dataset}

\item{vars}{The index variables in your data frame that identify (1) subjects, (2) timepoints, (3) trials (only for single trial), and (4) the responses. Insert the names as a vector in the order specified before (ie. c = ("subj", "time", "trial", "resp")), regardless of the order they occur in your dataset.}

\item{level}{Specify the level of your data, e.g. "single_trial", "ss" or "ga"}

\item{aggregate}{Shall single trial data be averaged to "ss" and "ga"?}
}
\value{
bosc-Object
}
\description{
\code{bosc_import}
Imports a dense sampling study's data frame into boscillator.
}
