% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bosc_import_surrogates.R
\name{bosc_import_surrogates}
\alias{bosc_import_surrogates}
\title{bosc_import}
\usage{
bosc_import_surrogates(
  bosc,
  data,
  sfreq,
  n_timepoints,
  vars = c("subj", "time", "n_surr", "hr"),
  level = "ss",
  aggregate = TRUE,
  overwrite = F
)
}
\arguments{
\item{bosc}{A bosc-object}

\item{data}{A dataframe with dense sampling data}

\item{sfreq}{in Hz}

\item{n_timepoints}{Number of unique timepoints in dataset}

\item{vars}{The index variables in your data frame that identify, for single trials, (1) subjects, (2) timepoints, (3) trials (only for single trial), (4) the responses and (5) number of surrogate dataset, or for single subject data, (1) subjects, (2) timepoints,  (3) number of surrogate dataset, and (4) the hitrate (or aggregated measure). Insert the names as a vector in the order specified before (e.g.. c = ("subj", "time", "n_surr", "hr")), regardless of the order they occur in your dataset.}

\item{level}{Specify the level of your imported data, e.g. "single_trial", "ss" or "ga"}

\item{aggregate}{Shall single trial data be averaged to "ss" and "ga"?}

\item{overwrite}{Argument passed to bosc_aggregate, defaults to F}
}
\value{
bosc-Object
}
\description{
\code{bosc_import}
}
