% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bosc_sim_default.R
\name{bosc_sim_default}
\alias{bosc_sim_default}
\title{bosc_sim_default}
\usage{
bosc_sim_default(
  surrogates = TRUE,
  detrend = TRUE,
  window = TRUE,
  scale = FALSE,
  pad = FALSE,
  fft = FALSE,
  sinmod = FALSE,
  fft_test = FALSE,
  sinmod_test = FALSE
)
}
\arguments{
\item{surrogates}{want to generate surrogates?}

\item{detrend}{want to detrend?}

\item{window}{want to window?}

\item{scale}{want to scale?}

\item{pad}{want to pad?}

\item{fft}{want to fft?}

\item{sinmod}{want sin modelling?}

\item{fft_test}{want stat test for fft?}

\item{sinmod_test}{want stat test for modelling?}
}
\value{
A BOSC-Object
}
\description{
\code{bosc_sim_default}
Wrapper function for bosc to simulate, preprocess and statistically test an experiment with default values.
}
\examples{
bosc <- bosc_sim_default()

}
