% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend_bosc.R
\name{detrend_bosc}
\alias{detrend_bosc}
\title{detrend_bosc}
\usage{
detrend_bosc(
  bosc,
  types = c("real", "surrogate"),
  levels = c("ss", "ga"),
  order = 0,
  verbose = TRUE
)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{types}{Which data should be detrended? choose between "real" and "surrogate" or concatenate to "real-surrogate" to perform aggregation on both.}

\item{levels}{Which levels of data need to be detrended? Use "ss" for single subject and "ga" for grand average. Concatenate multiple levels with "-", e.g. "ss-ga"}

\item{order}{Choose an order of the polynomial fit which will then be removed (0 = demean). Order = -1 will lead to skipping detrending.}

\item{verbose}{defaults to T}
}
\value{
A BOSC-Object
}
\description{
\code{detrend_bosc}
Detrends time courses in an BOSC-Object.
}
\examples{
bosc = simulate_experiment()
bosc = detrend_bosc(bosc, types = "real", levels = "ss", order = 1)

}
