% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_bosc.R
\name{fft_bosc}
\alias{fft_bosc}
\title{fft_bosc}
\usage{
fft_bosc(
  bosc,
  types = c("real", "surrogate"),
  levels = c("ss", "ga"),
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{types}{Which data should be FFTed? choose between "real" and "surrogate" or concatenate to "real-surrogate" to perform aggregation on both.}

\item{levels}{Which levels of data need to be FFTed? Use "ss" for single subject and "ga" for grand average. Concatenate multiple levels with "-", e.g. "ss-ga"}

\item{overwrite}{defaults to F}

\item{verbose}{defaults to T}
}
\value{
A BOSC-Object
}
\description{
\code{fft_bosc}
Performs an FFT on a time courses in an BOSC-Object.
}
\examples{
bosc = simulate_experiment()
bosc = fft_bosc(bosc, types = "real", levels = "ga")

}
