% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_bosc.R
\name{pad_bosc}
\alias{pad_bosc}
\title{pad_bosc}
\usage{
pad_bosc(
  bosc,
  types = "real-surrogate",
  levels = "ss-ga",
  method = "zero",
  n_pads = length(bosc$timepoints)
)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{types}{Which data should be padded? choose between "real" and "surrogate" or concatenate to "real-surrogate" to perform aggregation on both.}

\item{levels}{Which levels of data need to be padded? Use "ss" for single subject and "ga" for grand average. Concatenate multiple levels with "-", e.g. "ss-ga"}

\item{method}{Padding method? "zero" or "mean"}

\item{n_pads}{Number of pads}
}
\value{
A BOSC-Object
}
\description{
\code{pad_bosc}
Adds pads to the time courses in an BOSC-Object.
}
\examples{
bosc = simulate_experiment()
bosc = pad_bosc(bosc, types = "real", levels = "ga", method = "zero")

}
