% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_complex.R
\name{plot_complex}
\alias{plot_complex}
\title{plot_complex}
\usage{
plot_complex(bosc, freqs = unique(bosc$tests$fft$ss$complex$data$f))
}
\arguments{
\item{bosc}{BOSC-Object}

\item{freqs}{A vector of frequencies (in Hz) to plot, defaults to all frequencies available}
}
\value{
A BOSC-Object
}
\description{
\code{plot_complex}
Plots the results of an FFT complex test.
}
