% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fft_2d.R
\name{plot_fft_2d}
\alias{plot_fft_2d}
\title{plot_fft_2d}
\usage{
plot_fft_2d(bosc, freqs = unique(bosc$tests$fft$ss$complex$data$f))
}
\arguments{
\item{bosc}{BOSC-Object}

\item{freqs}{A vector of frequencies (in Hz) to plot, defaults to all frequencies available}
}
\value{
A BOSC-Object
}
\description{
\code{plot_fft_2d}
Plots the results of an FFT 2D test.
}
