% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phase.R
\name{plot_phase}
\alias{plot_phase}
\title{plot_phase}
\usage{
plot_phase(bosc, freqs = unique(bosc$data$ss$real$fft$f))
}
\arguments{
\item{bosc}{BOSC-Object}

\item{freqs}{A vector of frequencies (in Hz) to plot, defaults to all frequencies available}
}
\value{
A BOSC-Object
}
\description{
\code{plot_phase}
Plots the phase distributions
}
