% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{plot_spectrum}
\usage{
plot_spectrum(bosc, levels = "ss-merged_spectra-ga")
}
\arguments{
\item{bosc}{BOSC-Object}

\item{levels}{Which level shall be plotted? "ss", "ga" or "merged_spectra"}
}
\value{
A BOSC-Object
}
\description{
\code{plot_spectrum}
Plots the FFT frequency spectrum.
}
