% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_bosc.R
\name{scale_bosc}
\alias{scale_bosc}
\title{scale_bosc}
\usage{
scale_bosc(
  bosc,
  types = "real-surrogate",
  levels = "ss-ga",
  method = "z",
  verbose = T
)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{types}{Which data should be scaled? choose between "real" and "surrogate" or concatenate to "real-surrogate" to perform aggregation on both.}

\item{levels}{Which levels of data need to be scaled? Use "ss" for single subject and "ga" for grand average. Concatenate multiple levels with "-", e.g. "ss-ga"}

\item{method}{defaults to z}

\item{verbose}{defaults to T}
}
\value{
A BOSC-Object
}
\description{
\code{scale_bosc}
z-scales time courses in an BOSC-Object.
}
\examples{
bosc = simulate_experiment()
bosc = scale_bosc(bosc, types = "real", levels = "ga")

}
