% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_experiment.R
\name{simulate_experiment}
\alias{simulate_experiment}
\title{simulate_experiment}
\usage{
simulate_experiment(n_samples = 0)
}
\arguments{
\item{n_samples}{Number of samples}
}
\value{
A data frame
}
\description{
\code{simulate_experiment}
Simulates the raw dataset of a dense sampling study.
}
\examples{
data = simulate_experiment()

}
\author{
René Michel
}
