% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_time_course.R
\name{simulate_time_course}
\alias{simulate_time_course}
\title{simulate_time_course}
\usage{
simulate_time_course(n_samples = 0)
}
\arguments{
\item{n_samples}{Number of samples}
}
\value{
A vector
}
\description{
\code{simulate_time_course}
Simulates a time course from a dense sampling study.
}
\examples{
tc = simulate_time_course()

}
\author{
René Michel
}
