% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinmod_bosc.R
\name{sinmod_bosc}
\alias{sinmod_bosc}
\title{sinmod_bosc}
\usage{
sinmod_bosc(
  bosc,
  types = "real-surrogate",
  levels = "ss-ga",
  fixed_f = NULL,
  niter = 500,
  overwrite = FALSE,
  verbose = T
)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{types}{Which data should be modelled? choose between "real" and "surrogate" or concatenate to "real-surrogate" to perform aggregation on both.}

\item{levels}{Which levels of data need to be modelled? Use "ss" for single subject and "ga" for grand average. Concatenate multiple levels with "-", e.g. "ss-ga"}

\item{fixed_f}{NULL to test all possible frequencies between 0 and nyquist, or specify a vector of frequencies}

\item{overwrite}{defaults to F}

\item{verbose}{defaults to T}

\item{return_all}{only relevant if fixed_f is =! NULL. defaults to F. if T, results for all fixed frequencies will be returned. otherwise, only winner frequency results will be returned}
}
\value{
A BOSC-Object
}
\description{
\code{sinmod_bosc}
Fits sinusoidal model to a time courses in an BOSC-Object.
}
\examples{
bosc = simulate_experiment()
bosc = sinmod_bosc(bosc, types = "real", levels = "ga")

}
