% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_bosc.R
\name{window_bosc}
\alias{window_bosc}
\title{window_bosc}
\usage{
window_bosc(
  bosc,
  types = c("real", "surrogate"),
  levels = c("ss", "ga"),
  method = "hann",
  r = 0.1,
  alpha = 0.54,
  verbose = T
)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{types}{Which data should be windowed? choose between "real" and "surrogate" or concatenate to "real-surrogate" to perform aggregation on both.}

\item{levels}{Which levels of data need to be windowed? Use "ss" for single subject and "ga" for grand average. Concatenate multiple levels with "-", e.g. "ss-ga"}

\item{method}{Which window function should be used? defaults to "hann" for a hanning window}

\item{r}{Additional parameter for the tukey window function}

\item{alpha}{Additional parameter for the hamming, cosine and kaiser window function}

\item{verbose}{defaults to T}
}
\value{
A BOSC-Object
}
\description{
\code{window_bosc}
Applies a window function to the time courses in an BOSC-Object.
}
\examples{
bosc = simulate_experiment()
bosc = window_bosc(bosc, types = "real", levels = "ga", method = "hann")

}
