% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowing_bosc.R
\name{windowing_bosc}
\alias{windowing_bosc}
\title{windowing_bosc}
\usage{
windowing_bosc(
  bosc,
  types = "real-surrogate",
  levels = "ss-ga",
  method = "hann",
  r = 0.1,
  alpha = 0.54
)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{types}{Which data should be windowed? choose between "real" and "surrogate" or concatenate to "real-surrogate" to perform aggregation on both.}

\item{levels}{Which levels of data need to be windowed? Use "ss" for single subject and "ga" for grand average. Concatenate multiple levels with "-", e.g. "ss-ga"}

\item{method}{Which window function should be used? defaults to "hann" for a hanning window}

\item{r}{Additional parameter for the tukey window function}

\item{alpha}{Additional parameter for the hamming, cosine and kaiser window function}
}
\value{
A BOSC-Object
}
\description{
\code{windowing_bosc}
Applies a window function to the time courses in an BOSC-Object.
}
\examples{
bosc = simulate_experiment()
bosc = windowing_bosc(bosc, types = "real", levels = "ga", method = "hann")

}
