% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bosc_sim_default.R
\name{bosc_sim_default}
\alias{bosc_sim_default}
\title{bosc_sim_default}
\usage{
bosc_sim_default(surrogates = T, detrend = T, window = T, scale = F, pad = F)
}
\arguments{
\item{surrogates}{want to generate surrogates?}

\item{detrend}{want to detrend?}

\item{window}{want to window?}

\item{scale}{want to scale?}

\item{pad}{want to pad?}
}
\value{
A BOSC-Object
}
\description{
\code{bosc_sim_default}
Simulates and preprocesses an experiment with default values.
}
\examples{
bosc <- bosc_sim_default()

}
