% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_surrogates.R
\name{generate_surrogates}
\alias{generate_surrogates}
\title{generate_surrogates}
\usage{
generate_surrogates(
  bosc,
  n_surr = 100,
  method = "perm",
  n_seed = 1,
  overwrite = F,
  aggregate = T
)
}
\arguments{
\item{aggregate}{}
}
\value{
A list
}
\description{
\code{generate_surrogates}
Generates surrogate datasets from a dense sampling study
}
\examples{
bosc = simulate_experiment(n = 10, n_timepoints = 10, n_trials = 10)
bosc = generate_surrogates(bosc, n_surr = 100, method = "perm")

#' @author René Michel
}
