% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_surrogates.R
\name{generate_surrogates}
\alias{generate_surrogates}
\title{generate_surrogates}
\usage{
generate_surrogates(
  bosc,
  n_surr = 100,
  method = "perm",
  seed_num = NULL,
  overwrite = FALSE,
  aggregate = TRUE
)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{n_surr}{Number of to-be-created surrogate datasets}

\item{method}{permutation ("perm") or autoregression-model, either on single subject level only  (and averaged afterward; "ar_ss") or modelling both on single subject and grand average level ("ar_ss_ga")}

\item{seed_num}{number of the seed}

\item{overwrite}{overwrite existing surrogates? defaults to F}

\item{aggregate}{aggregate surrogated datasets? defaults to T}
}
\value{
A list
}
\description{
\code{generate_surrogates}
Generates surrogate datasets from a dense sampling study
}
\examples{
bosc = simulate_experiment(n = 10, n_timepoints = 10, n_trials = 10)
bosc = generate_surrogates(bosc, n_surr = 100, method = "perm")

}
\author{
René Michel
}
