% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_experiment.R
\name{simulate_experiment}
\alias{simulate_experiment}
\title{simulate_experiment}
\usage{
simulate_experiment(
  n_sub = 14,
  n_timepoints = 20,
  n_trials = 100,
  sfreq = 25,
  intercept = 0.5,
  amplitude = 0.1,
  freq = 4,
  phi = 0,
  phase_jitter_within_subj = 0,
  phase_jitter_across_subj = 0,
  amplitude_jitter_within_subj = 0,
  amplitude_jitter_across_subj = 0,
  freq_jitter_within_subj = 0,
  freq_jitter_across_subj = 0,
  intercept_jitter_across_subj = 0,
  transient = "none",
  transient_expModel_params = c(0, 1, 0.3),
  trend = "none",
  trend_linModel_params = c(intercept, amplitude),
  trend_expModel_params = c(0, 1 - 2 * amplitude, 0.6),
  aggregate = T,
  seed_num = NULL
)
}
\arguments{
\item{n_sub}{Number of subjects}

\item{n_timepoints}{Number of timepoints, e.g. number of SOAs probed}

\item{n_trials}{Number of trials per timepoint and participant}

\item{sfreq}{Sampling frequency of timepoints (in Hz)}

\item{intercept}{mean performance of the subject}

\item{amplitude}{amplitude of the underlying oscillation}

\item{freq}{frequency of the underlying oscillation}

\item{phi}{phase of the underlying oscillation (0 to 2pi)}

\item{phase_jitter_within_subj}{sd of phase distribution across trials within subjects}

\item{phase_jitter_across_subj}{sd of phase distribution across subjects}

\item{amplitude_jitter_within_subj}{sd of amp distribution across trials within subjects}

\item{amplitude_jitter_across_subj}{sd of amp distribution across subjects}

\item{freq_jitter_within_subj}{sd of freq distribution across trials within subjects}

\item{freq_jitter_across_subj}{sd of freq distribution across subjects}

\item{intercept_jitter_across_subj}{sd of mean performance across subjects}

\item{transient}{"none", "hanning", or "exponential"}

\item{transient_expModel_params}{model parameters for exponential transient}

\item{trend}{"none", "linear" or "exponential"}

\item{trend_linModel_params}{model parameters for linear trend in performance}

\item{trend_expModel_params}{model parameters for exponential trend in performance}

\item{aggregate}{aggregate single trial data to single subject and grand average time series? defaults to T}

\item{seed_num}{seed number. if not specified, a random seed will be used and saved alongside with the data.}
}
\value{
A data frame
}
\description{
\code{simulate_experiment}
Simulates the raw dataset of a dense sampling study.
}
\examples{
bosc = simulate_experiment()

}
\author{
René Michel
}
