% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_time_course.R
\name{simulate_time_course}
\alias{simulate_time_course}
\title{simulate_time_course}
\usage{
simulate_time_course(
  t = NULL,
  n_samples = 20,
  sfreq = 24,
  start = 0.192,
  seed_num = 1,
  snratio = 1,
  intercept = 0.5,
  amplitude = 0.1,
  frequency = 4,
  phi = 0
)
}
\arguments{
\item{n_samples}{Number of samples}
}
\value{
A vector
}
\description{
\code{simulate_time_course}
Simulates a time course from a dense sampling study.
}
\examples{
tc = simulate_time_course()

}
\author{
René Michel
}
