% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinmod_bosc.R
\name{sinmod_bosc}
\alias{sinmod_bosc}
\title{sinmod_bosc}
\usage{
sinmod_bosc(
  bosc,
  types = c("real", "surrogate"),
  levels = c("ss", "ga"),
  fixed_f = NULL,
  niter = 100,
  convergence_count = 50,
  supp_errors = "Y",
  overwrite = FALSE,
  verbose = T
)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{types}{Which data should be modelled? choose between "real" and "surrogate" or concatenate to "real-surrogate" to perform aggregation on both.}

\item{levels}{Which levels of data need to be modelled? Use "ss" for single subject and "ga" for grand average. Multiple arguments need to be passed as a vector.}

\item{fixed_f}{NULL to test all possible frequencies between 0 and nyquist, or specify a vector of frequencies}

\item{niter}{number of iterations for fitting, is piped into iter argument in nls.multstart function}

\item{convergence_count}{number of iterations until model is declared winner model, is piped into iter argument in nls.multstart function}

\item{supp_errors}{defaults to 'Y', suppresses error messages from nls.multstart?}

\item{overwrite}{defaults to F}

\item{verbose}{defaults to T}
}
\value{
A BOSC-Object
}
\description{
\code{sinmod_bosc}
Fits sinusoidal model to a time courses in an BOSC-Object using nls.multstart. If not specified otherwise in fixed_f, only the winner model according to nls.multstart (across all frequencies tested by the algorithm) will be reported.
}
\examples{
bosc = simulate_experiment()
bosc = sinmod_bosc(bosc, types = "real", levels = "ga")

}
