% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinmod_bosc.R
\name{sinmod_bosc}
\alias{sinmod_bosc}
\title{sinmod_bosc}
\usage{
sinmod_bosc(
  bosc,
  types = c("real", "surrogate"),
  levels = c("ss", "ga"),
  fixed_f = NULL,
  niter = NULL,
  grid = 3,
  convergence_count = FALSE,
  supp_errors = "Y",
  lower = list(intercept = 0, a = 0, f = NULL, phi = 0),
  upper = list(intercept = 1, a = 1, f = NULL, phi = 2 * pi),
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{types}{Which data should be modelled? choose between "real" and "surrogate" or concatenate to "real-surrogate" to perform aggregation on both.}

\item{levels}{Which levels of data need to be modelled? Use "ss" for single subject and "ga" for grand average. Multiple arguments need to be passed as a vector.}

\item{fixed_f}{NULL to test all possible frequencies between 0 and nyquist, or specify a vector of frequencies}

\item{niter}{number of iterations for fitting, is piped into iter argument in nls.multstart function if NULL, a grid search approach will be used}

\item{grid}{number of start values per parameter, will be equally distributed within parameter boundaries, e.g. grid = 5 will create a 5x5x5x... grid, argument will be ignored if iter != NULL, if grid = NULL and iter = NULL, the grid will automatically be generated with grid = ceiling((nyquist-fres)/2)}

\item{convergence_count}{number of iterations until model is declared winner model, is piped into iter argument in nls.multstart function, will be ignored if grid search approach is used}

\item{supp_errors}{defaults to 'Y', suppresses error messages from nls.multstart?}

\item{lower}{lower boundaries for model estimates. if NULL is chosen for f, it will automatically be set to the lowest possible frequency given the frequency resolution in the dataset}

\item{upper}{upper boundaries for model estimates. if NULL is chosen for f, it will automatically be set to the nyquist frequency in the dataset}

\item{overwrite}{defaults to F}

\item{verbose}{defaults to T}
}
\value{
A BOSC-Object
}
\description{
\code{sinmod_bosc}
Fits sinusoidal model to a time courses in an BOSC-Object using nls.multstart. If not specified otherwise in fixed_f, only the winner model according to nls.multstart (across all frequencies tested by the algorithm) will be reported.
}
\examples{
bosc = simulate_experiment()
bosc = sinmod_bosc(bosc, types = "real", levels = "ga")

}
