% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fft.R
\name{test_fft}
\alias{test_fft}
\title{test_fft}
\usage{
test_fft(
  bosc,
  levels = c("ss", "merged", "ga"),
  tests = c("amp", "complex", "phase"),
  alpha = 0.05,
  mcc = c("bonferroni", "fdr", "maxfreq"),
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{levels}{Which levels of data need to be padded? Use "ss" for single subject and "ga" for grand average. Concatenate multiple levels with "-", e.g. "ss-ga"}

\item{tests}{Which test to perform?}

\item{alpha}{Vector of alpha levels to apply. If you want to use multiple alpha levels, make sure to set mcc to "none"}

\item{mcc}{choose between all methods available for stats::p.adjust, e.g."bonferroni", "fdr", and "maxfreq" or "none". if != "none", the first entry in alpha will be considered the family wise alpha error and all other entries in alpha will be ignored}

\item{overwrite}{defaults to F}

\item{verbose}{defaults to T}
}
\value{
A BOSC-Object
}
\description{
\code{test_fft}
Performs statistical test on frequency spectra obtained from fft_bosc.
}
\examples{
bosc = simulate_experiment()
bosc = generate_surrogates(bosc)
bosc = fft_bosc(bosc)
bosc = test_fft(bosc, levels = "ga", tests = "amp-complex")

}
