% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fft.R
\name{test_fft}
\alias{test_fft}
\title{test_fft}
\usage{
test_fft(
  bosc,
  levels = "ss-merged_spectra-ga",
  tests = "amp-complex-phase",
  alpha_amp = 0.05,
  alpha_complex = alpha_amp,
  alpha_phase = alpha_amp,
  overwrite = FALSE,
  verbose = T
)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{levels}{Which levels of data need to be padded? Use "ss" for single subject and "ga" for grand average. Concatenate multiple levels with "-", e.g. "ss-ga"}

\item{tests}{Which test to perform?}

\item{alpha_amp}{Which alpha level to apply for amplitude test}

\item{alpha_complex}{Which alpha level to apply for complex plane test}

\item{alpha_phase}{Which alpha level to apply for rayleigh test}

\item{overwrite}{defaults to F}

\item{verbose}{defaults to T}
}
\value{
A BOSC-Object
}
\description{
\code{test_fft}
Performs an FFT on a time courses in an BOSC-Object.
}
\examples{
bosc = simulate_experiment()
bosc = generate_surrogates(bosc)
bosc = fft_bosc(bosc)
bosc = test_fft(bosc, levels = "ga", tests = "amp-complex")

}
