% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fft.R
\name{test_fft}
\alias{test_fft}
\title{test_fft}
\usage{
test_fft(bosc, levels = "ss-ga", alpha = 0.05, overwrite = FALSE)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{levels}{Which levels of data need to be padded? Use "ss" for single subject and "ga" for grand average. Concatenate multiple levels with "-", e.g. "ss-ga"}

\item{alpha}{Which alpha level to apply?}

\item{overwrite}{defaults to F}
}
\value{
A BOSC-Object
}
\description{
\code{test_fft}
Performs an FFT on a time courses in an BOSC-Object.
}
\examples{
bosc = simulate_experiment()
bosc = generate_surrogates(bosc)
bosc = fft_bosc(bosc)
bosc = test_fft(bosc, levels = "ga")

}
