% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sinmod.R
\name{test_sinmod}
\alias{test_sinmod}
\title{test_sinmod}
\usage{
test_sinmod(
  bosc,
  levels = c("ss", "ga"),
  tests = c("r2"),
  alpha = 0.05,
  mcc = c("bonferroni", "fdr"),
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{bosc}{BOSC-Object}

\item{levels}{Which levels of data need to be padded? Use "ss" for single subject and "ga" for grand average. Concatenate multiple levels with "-", e.g. "ss-ga"}

\item{tests}{Which test to perform?}

\item{alpha}{Vector of alpha levels to apply. If you want to use multiple alpha levels, make sure to set mcc to "none"}

\item{mcc}{choose between "bonferroni", "fdr" and "none". if != "none", the first entry in alpha will be considered the family wise alpha error and all other entries in alpha will be ignored}

\item{overwrite}{defaults to F}

\item{verbose}{defaults to T}
}
\value{
A BOSC-Object
}
\description{
\code{test_sinmod}
Performs statistical test on sinusoidal models obtained from sinmod_bosc.
}
\examples{
bosc = simulate_experiment(n_sub = 3)
bosc = generate_surrogates(bosc, n_surr = 10)
bosc = sinmod_bosc(bosc)
bosc = test_sinmod(bosc, levels = "ga", tests = "r2")

}
